More building features
-Теперь рабочие могут хилить здания на кнопку R(нужно подойти вплотную к зданию и иметь свободное действие)
-Теперь поставить новое строение можно только рядом с союзным рабочим
-Теперь рабочий может выходить из здания(сделано частично)

+++ Assets/Scripts/UnitActions.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class UnitActions : MonoBehaviour
 {
-    public byte remainingActionsCount;
+    public int remainingActionsCount;
     private HexGrid _hexGrid => FindObjectOfType<HexGrid>();
     private SelectionController _selectionController => FindObjectOfType<SelectionController>();
     private HighlightingController _highlightedController => FindObjectOfType<HighlightingController>();
