+++ Assets/Scripts/UnitMovement.cs
@@ -13,8 +13,8 @@ public class UnitMovement : MonoBehaviour
     private bool _isHighlightedNeighbour;
     private ObjectOnGrid _objectOnGrid => GetComponent<ObjectOnGrid>();
     private PlacementManager _placementManager => FindObjectOfType<PlacementManager>();
-    private short _maxSpeed;
-    public short spentSpeed = 0;
+    private int _maxSpeed;
+    public int spentSpeed = 0;
     private Transform _highlighted;
     private MouseSelection _mouseSelection => FindObjectOfType<MouseSelection>();
     private HexGrid _hexGrid => FindObjectOfType<HexGrid>();
